Convert all tests to Catch

+++ src/lib/neo/fft/transform/radix2.hpp
@@ -51,7 +51,7 @@ auto c2c_radix2(std::span<Complex, Extent> x, Range const& twiddles) -> void
 {
     // bit-reverse ordering
     auto const len   = x.size();
-    auto const order = ilog2(len);
+    auto const order = static_cast<std::int32_t>(ilog2(len));
     bit_reverse_permutation(x);
 
     // butterfly computation
@@ -65,10 +65,10 @@ auto c2c_radix2(std::span<Complex, Extent> x, Range const& twiddles) -> void
 
         for (auto k = 0; std::cmp_less(k, len); k += stride) {
             for (auto pair = 0; pair < stage_length; ++pair) {
-                auto const tw = twiddles[pair * tw_stride];
+                auto const tw = twiddles[static_cast<std::size_t>(pair * tw_stride)];
 
-                auto const i1 = k + pair;
-                auto const i2 = k + pair + stage_length;
+                auto const i1 = static_cast<std::size_t>(k + pair);
+                auto const i2 = static_cast<std::size_t>(k + pair + stage_length);
 
                 auto const temp = x[i1] + tw * x[i2];
                 x[i2]           = x[i1] - tw * x[i2];
